--- junit5/README.md
@@ -1,2 +1,0 @@
-## Relevant articles: 
-- [The Order of Tests in JUnit](http://www.baeldung.com/junit-5-test-order)
